refactor loading config to config module
