+++ src/main/python/afp_cli/cli.py
@@ -29,7 +29,6 @@ import socket
 import subprocess
 import sys
 import tempfile
-import yamlreader
 
 from docopt import docopt
 from afp_cli import AWSFederationClientCmd, aws_credentials_file
@@ -37,22 +36,7 @@ import afp_cli.cli_functions as cli
 from afp_cli.log import error, debug
 from .password_providers import get_password
 from . import log
-
-CFGDIR = '/etc/afp-cli'
-
-
-def load_config(global_config_dir=CFGDIR):
-    global_config = {}
-    if os.path.isdir(global_config_dir):
-        global_config = yamlreader.yaml_load(global_config_dir, {})
-
-    user_config = {}
-    user_config_dir = os.path.expanduser("~/.afp-cli")
-    if os.path.isdir(user_config_dir):
-        global_config = yamlreader.yaml_load(user_config_dir, {})
-
-    yamlreader.data_merge(global_config, user_config)
-    return global_config
+from .config import load_config
 
 
 def get_default_afp_server():
